Heavy Backend cleanup from bullshit LLM-generated code logic
